Improve anagrams
The problem of anagrams are no more! Now, we will by default remove all anagrams and then when we go to add a word, we will take all the combinations of anagrams and non anagram to get all of the possibilities without expending the resources to find them in the wild.

So we end up with some domesticated anagrams, or, as the scientists may call them, domesticada samgranate