Crazy Improvement
So I used the "must have a letter" technique to shave off a large portion of words. I then improved on this technique.

There is a new option: `guessoffset` which stipulates how many extra seed letters we will consider above the minimum. If we add an extra letter, then for all words with only 1 of those letters, it will force the next guess to get a word with one of those problem letters, and since our data is now sorted according to how many problem letters that word has, we basically start off at ~90% completion, and many of the words with only 1 problem letter with an offset of 1 have a majority of their neighbors discarded since they need at minimum 2/3 of those problem letters (in the 5x5)